requirements.txt
# Core ML dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data validation
pandera>=0.17.0

# API serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
pyyaml>=6.0
python-multipart>=0.0.6
httpx>=0.24.0
structlog>=23.0.0